Add support for direct prometheus queries, OOM events
There are issues with using external-metrics, namely only one
external-metrics / custom-metrics provider is allowed per cluster.
Popular component KEDA also takes the external-metrics path which causes
these two components to be incompatible if we run VPA in
external-metrics mode.
There is also an issue with managed memory languages that do not allow
themselves to be OOMKilled(Such as .NET) will not get the OOMKilled bump
when they need, so we add support for detecting these events via
Prometheus.